Rename Routing\LegacyRouteParser as Routing\Legacy\RouteParser
